Fix dependency vulnerabilities
The versions of axios and eslint we're depending on have known security
vulnerabilities, and since we've locked the versions, NPM will dutifully
install the vulnerable versions for our users. That's obviously not
acceptable.

This commit updates to the latest versions, but also relaxes the version
lock to allow PATCH and MINOR version changes.